代码优化

+++ BeautifulTimes/Src/IMModule/ChatModel/View/BTSendTextView.m
@@ -28,20 +28,19 @@
 
 - (void)appendEmotion:(HMEmotion *)emotion
 {
-    if (emotion.emoji) { // emoji表情
+    if (emotion.emoji) {
         [self insertText:emotion.emoji];
-    } else { // 图片表情
-        //[self insertText:emotion.chs];
+    }
+    else {
         NSMutableAttributedString *attributedText = [[NSMutableAttributedString alloc] initWithAttributedString:self.attributedText];
         
-        // 创建一个带有图片表情的富文本
         HMEmotionAttachment *attach = [[HMEmotionAttachment alloc] init];
         attach.emotion = emotion;
         attach.bounds = CGRectMake(0, -3, self.font.lineHeight, self.font.lineHeight);
         NSAttributedString *attachString = [NSAttributedString attributedStringWithAttachment:attach];
         
         // 记录表情的插入位置
-        int insertIndex = self.selectedRange.location;
+        long insertIndex = self.selectedRange.location;
         
         // 插入表情图片到光标位置
         [attributedText insertAttributedString:attachString atIndex:insertIndex];
@@ -59,21 +58,18 @@
 
 - (NSString *)messageText
 {
-    // 1.用来拼接所有文字
     NSMutableString *string = [NSMutableString string];
     
-    // 2.遍历富文本里面的所有内容
     [self.attributedText enumerateAttributesInRange:NSMakeRange(0, self.attributedText.length) options:0 usingBlock:^(NSDictionary *attrs, NSRange range, BOOL *stop) {
         HMEmotionAttachment *attach = attrs[@"NSAttachment"];
-        if (attach) { // 如果是带有附件的富文本
+        if (attach) {
             [string appendString:attach.emotion.chs];
-        } else { // 普通的文本
-            // 截取range范围的普通文本
+        }
+        else {
             NSString *substr = [self.attributedText attributedSubstringFromRange:range].string;
             [string appendString:substr];
         }
     }];
-    
     return string;
 }
